Small reorganization into smaller modules in preparation for #18
